[fix] int -> uint narrowing conversion
- int -> uint로의 축소 변환 컴파일 에러 수정
- chunkSize는 최대 12,000(MTU << 3)을 넘을 수 없으므로 uint16_t 사용하도록 변경

+++ server/Src/DabaebmintonServerCore/RUDPHeader.cpp
@@ -5,13 +5,19 @@
 #include "Constant.h"
 #include "GetRequiredBits.h"
 
-RUDPHeader::RUDPHeader(uint8_t channelId, uint8_t sequenceNumber, uint32_t totalBitLength, uint32_t offset, uint32_t chunkLength) :
+RUDPHeader::RUDPHeader(
+	uint8_t channelId,
+	uint8_t sequenceNumber,
+	uint32_t totalBitLength,
+	uint32_t offset, 
+	uint16_t chunkLength) :
 	_channelId(channelId),
 	_sequenceNumber(sequenceNumber),
 	_totalBitLength(totalBitLength),
 	_offset(offset),
 	_chunkLength(chunkLength)
-{}
+{
+}
 
 void RUDPHeader::Write(OutputMemoryBitStream& outStream) const
 {
